@@ -85,8 +85,7 @@
     "react-simple-dropdown": "^3.2.0",
     "react-tabs": "^2.2.2",
     "rxdb": "^7.6.1",
-    "rxjs": "^6.2.0",
-    "urijs": "^1.19.1"
+    "rxjs": "^6.2.0"
   },
   "devDependencies": {
     "babel-core": "^6.26.3",
@@ -94,15 +93,15 @@
     "babel-loader": "^7.1.4",
     "babel-plugin-istanbul": "^3.0.0",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "babel-plugin-transform-runtime": "^6.15.0",
     "babel-polyfill": "^6.26.0",
-    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
     "babel-runtime": "^6.26.0",
     "better-npm-run": "0.0.13",
     "compression": "^1.7.2",
-    "copy-webpack-plugin": "^4.4.1",
+    "copy-webpack-plugin": "^4.5.2",
     "css-loader": "^0.28.11",
     "cssnano": "^3.7.4",
     "debug": "^2.6.9",
@@ -110,7 +109,7 @@
     "eslint-config-airbnb": "^12.0.0",
     "eslint-import-resolver-webpack": "^0.8.4",
     "eslint-loader": "^2.0.0",
-    "eslint-plugin-import": "^2.0.1",
+    "eslint-plugin-import": "^2.13.0",
     "eslint-plugin-react": "^6.4.1",
     "express": "^4.16.3",
     "extract-text-webpack-plugin": "^4.0.0-beta.0",
@@ -126,9 +125,10 @@
     "sass-loader": "^7.0.3",
     "style-loader": "^0.21.0",
     "url-loader": "^1.0.1",
-    "webpack": "^4.0.0",
+    "urijs": "^1.19.1",
+    "webpack": "^4.12.2",
+    "webpack-cli": "^3.0.8",
     "webpack-bundle-analyzer": "^2.13.1",
-    "webpack-cli": "^3.0.3",
     "webpack-dev-middleware": "^3.1.3",
     "webpack-dev-server": "^3.1.4",
     "webpack-hot-middleware": "^2.21.0",
