@@ -3,7 +3,7 @@ import BEMHelper from 'react-bem-helper';
 import EmptyState from 'components/EmptyState';
 import './style.scss';
 
-function LoadingPanel({ message, isActive }) {
+function LoadingPanel({ isActive, message = 'Загрузка' }) {
 	const classes = new BEMHelper('loading-panel');
 	const loader = (<div>
 		<img src={'/loader.gif'} {...classes('image')} />
