@@ -6,14 +6,16 @@ import UserInfo from './UserInfo';
 import Select from './Select';
 import Input from './Input';
 import LoadingPanel from './LoadingPanel';
+import Modal from './Modal';
 
 export {
-	Button,
-	Header,
-	EmptyState,
-	UserInfo,
-	ListItem,
-	Select,
-	Input,
-	LoadingPanel,
+  Button,
+  Header,
+  EmptyState,
+  UserInfo,
+  ListItem,
+  Select,
+  Input,
+  LoadingPanel,
+  Modal,
 };
