@@ -1,3 +1,4 @@
+// @ts-check
 import React, { Component } from 'react';
 import { browserHistory, Router } from 'react-router';
 
@@ -12,7 +13,9 @@ class AppContainer extends Component {
 
     return (
       <div style={{ height: '100%' }}>
-        <Router history={browserHistory} children={routes} />
+        <Router history={browserHistory}>
+          {routes}
+        </Router>
       </div>
     );
   }
