@@ -1,13 +1,13 @@
+// @ts-check
 import React, { Component } from 'react';
 import BEMHelper from 'react-bem-helper';
-// import { connect } from 'react-redux';
 
-// import { Budget } from 'routes/Budget/containers';
-import { Budgets } from 'routes/Budgets/containers';
-import { Construct } from 'routes/Constructor/containers';
-import { Login } from 'routes/Login/containers';
+import { Budget } from '../../routes/Budget/containers';
+import { Budgets } from '../../routes/Budgets/containers';
+import { Construct } from '../../routes/Constructor/containers';
+import { Login } from '../../routes/Login/containers';
 // import { Collaborators } from 'routes/Collaborators/containers';
-import { notify } from 'services/helpers';
+import { notify } from '../../services/helpers';
 
 import './CoreLayout.scss';
 import { Database } from '../../database/index';
@@ -16,8 +16,8 @@ import { GlobalStore } from '../../store/globalStore';
 const MobileDetect = require('mobile-detect');
 
 export class CoreLayout extends Component {
-  constructor() {
-    super();
+  constructor(params) {
+    super(params);
     this.state = {
       isOffline: false,
       isLoggedIn: GlobalStore.modules.auth.isLoggedIn.value,
@@ -55,16 +55,16 @@ export class CoreLayout extends Component {
     const { isOffline, isLoggedIn } = this.state;
 
     return (
-      <div className="container">
+      <div {...classes()}>
         <div {...classes({ element: 'viewport', modifiers: { mobile: isMobile, offline: isOffline } })}>
           {isLoggedIn === true
             ? [
-              /*<Collaborators />,
-              <Budget />,*/
-              <Construct />,
-              <Budgets />,
+              /*<Collaborators key={1} />,*/
+              <Budget key={2} />,
+              <Construct key={3} />,
+              <Budgets key={4} />,
             ]
-            : <Login />
+            : <Login key={5} />
           }
         </div>
       </div>
