@@ -1,10 +1,11 @@
-import { Observable } from "../../providers/Observable";
+// @ts-check
+import { Observable } from '../../providers/Observable';
 import { actions as authActions } from './actions';
 
 const token = authActions.restoreToken();
 
 export default {
-  namespace: 'auth',
+  path: 'auth',
   state: {
     isLoggedIn: new Observable(!!token),
     token: new Observable(token),
