@@ -1,7 +1,10 @@
+// @ts-check
 import auth from './auth';
 import users from './users';
+import budgets from './budgets';
 
 export default [
-	auth,
-	users,
+  auth,
+  users,
+  budgets,
 ];
