@@ -1,6 +1,7 @@
-import { GlobalStore } from "../../store/globalStore";
+// @ts-check
+import { GlobalStore } from '../../store/globalStore';
 
-const setActiveUser = function(user) {
+const setActiveUser = function (user) {
   GlobalStore.modules.users.activeUser.value = user;
 };
 
