@@ -1,11 +1,12 @@
+// @ts-check
 import { Component } from 'react';
 import BEMHelper from 'react-bem-helper';
 import { GlobalStore } from '../store/globalStore';
 
 export class Page extends Component {
-  constructor() {
-    super();
-    this.namespace = '';
+  constructor(props) {
+    super(props);
+    this.path = '';
     this.state = {
       isActive: false,
       isNext: false,
@@ -22,29 +23,29 @@ export class Page extends Component {
 
   componentWillMount() {
     this.setState({
-      isActive: GlobalStore.routes.active.value === this.namespace,
-      isNext: GlobalStore.routes.next.value.includes(this.namespace),
-      isPrev: GlobalStore.routes.prev.value.includes(this.namespace),
+      isActive: GlobalStore.routes.active.value === this.path,
+      isNext: GlobalStore.routes.next.value.includes(this.path),
+      isPrev: GlobalStore.routes.prev.value.includes(this.path),
     });
   }
 
   onActiveRouteChange(route) {
     this.setState({
-      isActive: route === this.namespace,
+      isActive: route === this.path,
     });
   }
 
   onNextRoutesChange(nextRoutes) {
-      this.setState({
-      isNext: nextRoutes.includes(this.namespace),
-      });
-    }
+    this.setState({
+      isNext: nextRoutes.includes(this.path),
+    });
+  }
 
   onPrevRoutesChange(prevRoutes) {
-      this.setState({
-      isPrev: prevRoutes.includes(this.namespace),
-      });
-    }
+    this.setState({
+      isPrev: prevRoutes.includes(this.path),
+    });
+  }
 
   getPageClasses() {
     return this.bemHelper({
