@@ -1,39 +1,38 @@
-// import { connect } from 'react-redux';
-import { withRouter } from 'react-router';
-import Summary from './presenter';
+// @ts-check
+import { Component } from 'react';
+import presenter from './presenter';
+import { budgetsActions } from '../../../../modules/budgets/actions';
+import { GlobalStore } from '../../../../store/globalStore';
 
-function getBalance(total, my, count) {
-  if (count === 0) {
-    return 0;
+export class Summary extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      total: 0,
+      myTotal: 0,
+    };
   }
-  return my - total / count;
+
+  componentWillReceiveProps(nextProps) {
+    const {
+      transactions = [],
+      users = [],
+    } = nextProps;
+    const balances = budgetsActions.calculateBalances(transactions, users);
+    const total = transactions
+      .filter(transaction => !transaction.cancelled)
+      .reduce((prev, current) => prev + current.amount, 0);
+
+    this.setState({
+      total,
+      myTotal: balances.get(GlobalStore.modules.users.activeUser.value.id) || 0,
+    });
+  }
+
+  render() {
+    return presenter({
+      ...this.props,
+      ...this.state,
+    });
+  }
 }
-
-const mapDispatchToProps = {
-
-};
-
-const mapStateToProps = (state, ownProps) => {
-  let total = 0;
-  let myTotal = 0;
-  const {
-    transactions = [],
-    users = {},
-    currentUserId,
-  } = ownProps;
-  const activeUsers = Object.values(users).filter(user => user.status === 'active');
-  transactions.forEach((t) => {
-    total += !t.cancelled ? t.amount : 0;
-  });
-  transactions.forEach((t) => {
-    myTotal += t.ownerId === currentUserId && !t.cancelled ? t.amount : 0;
-  });
-  myTotal = getBalance(total, myTotal, activeUsers.length);
-
-  return {
-    total,
-    myTotal,
-  };
-};
-
-export default Summary;
