@@ -1,11 +1,10 @@
 import React from 'react';
 import BEMHelper from 'react-bem-helper';
-import { Modal } from 'components/Modal';
-import Button from 'components/Button';
-import { decisions } from 'routes/Budget/const';
-import UserInfo from '../UserInfo';
-
 import './style.scss';
+import { decisions } from '../../../const';
+import { Modal, Button } from '../../../../../components';
+import { GlobalStore } from '../../../../../store/globalStore';
+import { UserInfo } from '../UserInfo/presenter';
 
 function Icon(props) {
   const classes = new BEMHelper('modal-closing');
@@ -26,20 +25,20 @@ function Icon(props) {
 export default function ModalClosing(props) {
   const classes = new BEMHelper('modal-closing');
   const {
-    usersList = [],
-    currentUserId,
+    users = [],
     makeDecision,
+    ownerId,
   } = props;
-  const currentUser = usersList.find(user => user.id === currentUserId) || {};
+  const currentUserId = GlobalStore.modules.users.activeUser.value.id;
+  const currentUser = users.find(user => user.id === currentUserId) || {};
   const canMakeDecision = currentUser.decision === decisions.pending && !currentUser.isOwner;
-
   return (
     <Modal title={'Ждем подтверждения от участников'}>
       <ul>
-        {usersList
+        {users
           .map((user) => {
             let status = decisions.pending;
-            if (user.decision === decisions.approved || user.isOwner) {
+            if (user.decision === decisions.approved || user.id === ownerId) {
               status = decisions.approved;
             }
             return (<UserInfo {...user}>
