@@ -1,25 +1,23 @@
 import React from 'react';
 import BEMHelper from 'react-bem-helper';
-import get from 'lodash/get';
 
 import './style.scss';
 
-export default function UserInfo(props) {
-	const {
-		id,
+export function UserInfo(props) {
+  const {
     name,
     phone,
     email,
     children,
-	} = props;
-	const classes = new BEMHelper('budget-user-info');
+  } = props;
+  const classes = new BEMHelper('budget-user-info');
 
-	return (
-		<div {...classes()}>
-			<div {...classes('name')}>
+  return (
+    <div {...classes()}>
+      <div {...classes('name')}>
         {name || phone || email}
       </div>
       {children}
-		</div>
-	);
+    </div>
+  );
 }
