@@ -1,6 +1,6 @@
-const namespace = 'budgets';
+const path = '/budgets';
 const paths = {
-  budget: id => `/budgets/${id}`,
+  budget: id => `/budget/${id}`,
   construct: () => '/create',
 };
 const apiPaths = {
@@ -10,5 +10,5 @@ const apiPaths = {
 export {
   apiPaths,
   paths,
-  namespace,
+  path,
 };
