@@ -1,13 +1,14 @@
 import { Route } from '../../providers/Route';
-import { namespace } from './const';
-import { namespace as constructNamespace } from 'routes/Constructor/const';
+import { path } from './const';
+import { path as constructpath } from '../Constructor/const';
+import { path as budgetpath } from '../Budget/const';
 import { Budgets } from './containers';
 
 class BudgetsRoute extends Route {
   constructor() {
     super();
-    this.path = namespace;
-    this.nextRoutes = ['budget', constructNamespace];
+    this.route = path;
+    this.nextRoutes = [budgetpath, constructpath];
   }
 
   getContainer() {
