@@ -1,7 +1,11 @@
+// @ts-check
+const path = '/collaborators';
+
 const paths = {
-    budget: (id) => `/budgets/${id}`,
+  budget: (id) => `/budgets/${id}`,
 };
 
 export {
-    paths,
+  paths,
+  path,
 };
