@@ -1,14 +1,15 @@
+// @ts-check
 import { Route } from '../../providers/Route';
-import { namespace } from './const';
-import { namespace as budgetsNamespace } from 'routes/Budgets/const';
+import { path } from './const';
+import { path as budgetspath } from 'routes/Budgets/const';
 import { Construct } from './containers';
 
 class ConstructRoute extends Route {
   constructor() {
     super();
-    this.path = namespace;
+    this.route = path;
     this.nextRoutes = [];
-    this.prevRoutes = [budgetsNamespace];
+    this.prevRoutes = [budgetspath];
   }
 
   getContainer() {
