@@ -1,6 +1,6 @@
 // We only need to import the modules necessary for initial render
 import CoreLayout from '../layouts/CoreLayout';
-// import BudgetRoute from './Budget';
+import BudgetRoute from './Budget';
 import ConstructorRoute from './Constructor';
 import BudgetsRoute from './Budgets';
 import LoginRoute from './Login';
@@ -9,7 +9,7 @@ import LoginRoute from './Login';
 /*  Note: Instead of using JSX, we recommend using react-router
     PlainRoute objects to build route definitions.   */
 
-export const createRoutes = {
+export const routes = {
   path: '/',
   component: CoreLayout,
   getIndexRoute(partialNextState, callback) {
@@ -19,16 +19,17 @@ export const createRoutes = {
   getChildRoutes(partialNextState, callback) {
     const childRoutes = [
       BudgetsRoute,
-      // BudgetRoute,
+      BudgetRoute,
       LoginRoute,
       ConstructorRoute,
       // CollaboratorsRoute,
     ];
-    const path = partialNextState.location.pathname.substr(1);
-    const activeRoute = childRoutes.find(route => route.path === path);
+    
+    const path = partialNextState.location.pathname;
+    const activeRoute = childRoutes.find(route => route.pathMatcher.test(path));
     activeRoute.onEnter();
     callback(null, activeRoute);
   },
 };
 
-export default createRoutes;
+export default routes;
